#JRAB-20 dodavanje colone priority za Todo model

+++ app/Http/Controllers/TodosController.php
@@ -13,14 +13,22 @@ class TodosController extends Controller
     	return $todos;
     }
 
+    public function show($id) {
+        $todo = Todo::find($id);
+        return $todo;
+   }
+
     public function store(Request $request) {
     	$todo = new Todo();
     	// validate
     	$rules = Todo::STORE_RULES;
+        // $options = Todo::OPTIONS;
+        // dd($options);
     	$request->validate($rules);
     	// new data
     	$todo->title = $request->input('title');
         $todo->completed = $request->input('completed');
+        $todo->priority = $request->get('priority');
     	// check if todo already exists in db
     	$existingTodo = Todo::where('title', '=', Input::get('title'))->first();
     	// save new Todo()
@@ -38,12 +46,16 @@ class TodosController extends Controller
         }
         $todo->title = $request->input('title');
         $todo->completed = $request->input('completed');
+        $todo->priority = $request->get('priority');
         $todo->save();
         return $todo;
     }
 
     public function destroy($id) {
         $todo = Todo::find($id);
+        if(! $todo) {
+            return response()->json(['message' => 'Todo to delete not found'], 404);
+        }
         $todo->delete();
         return response()->json(['message' => 'Todo deleted'], 200);
     }
